requirements.txt
pyaml
python-binance
python-telegram-bot